backend/requirements.txt
flask
flask-cors
praw
nltk
vaderSentiment
pandas
numpy
python-dotenv
feedparser
newsapi-python
gunicorn
scikit-learn
tensorflow-cpu
scipy
pymongo
